visual-test-old/requirements.txt
selenium>=4.19.0
webdriver-manager>=4.0.1
pillow>=10.3.0
pytest>=8.4.1
pytest-html>=4.1.1
imageio>=2.34.0
